Moved card-list routes into authroutes folder and incorporated a user check on add to collection click
