+++ views/layouts/main.handlebars
@@ -16,6 +16,7 @@
     </header>
     <nav>
         {{!-- <a href="/signin">login</a> --}}
+        <a href="/card-list">Add cards</a>
         <a href="/collection">My collection</a>
         {{#if user}}
         <a href="/logout">Logout</a>
